// src/app/views/response-area/calibration-exam/calibration-exam-component/calibration-exam.interface.ts
import { CommonResponseAreaInterface } from "../../../../interfaces/page-definition.interface";


export interface CalibrationExamInterface extends CommonResponseAreaInterface {
    exportToCSV?: boolean;
    tabsintId?: string;
    frequencies?: number[];
    targetLevels?: number[];
    responseRequired?: boolean
    showResults?: boolean
}

export interface CalibrationResultViewerInterface extends CommonResponseAreaInterface {
    displayRightEar: boolean;
    displayLeftEar: boolean;
}